LPS-127106 Rename dragAndDrop directory to drag-and-drop
